Split io.h.
