+++ apps/qsim_base.cc
@@ -22,7 +22,7 @@
 #include "../lib/circuit_qsim_parser.h"
 #include "../lib/fuser_basic.h"
 #include "../lib/gates_qsim.h"
-#include "../lib/io.h"
+#include "../lib/io_file.h"
 #include "../lib/parfor.h"
 #include "../lib/run_qsim.h"
 #include "../lib/simmux.h"
@@ -100,8 +100,8 @@ int main(int argc, char* argv[]) {
   }
 
   Circuit<GateQSim<float>> circuit;
-  if (!CircuitQsimParser<IO>::FromFile(opt.maxtime, opt.circuit_file,
-                                       circuit)) {
+  if (!CircuitQsimParser<IOFile>::FromFile(opt.maxtime, opt.circuit_file,
+                                           circuit)) {
     return 1;
   }
